Using `tabula-extractor` to liberate tables from their PDF imprisonment
Extracting tables from a PDF and transforming them into a spreadsheet format is not simple. Until quite recently this was almost impossible, let alone accomplishing such tasks with Free Software utilities.

The reason is: PDF as a document format does not store the meaning and the context of text elements. (This may improve in the future with the proliferation of tagged PDF/UA ( https://en.wikipedia.org/wiki/PDF/UA ) ( UA == Universal Accessibility ).

This has changed now, with the arrival of tabula-extractor ( https://github.com/tabulapdf/tabula-extractor ) on the scene. That tool can extract tables which are trapped in a PDF's fixed layout even though their data may not be tagged at all as 'table' and 'colum' and 'row' and 'cell' .

This ASCiinema screencast introduces you to tabula-extractor ( https://github.com/tabulapdf/tabula-extractor ) . It covers installation from Git, preparing a small wrapper script and putting it to use with a 293 pages long PDF consisting of one loooong table...